requirements.txt
PyYAML
Faker
vertica-python
Jinja2
tqdm